[hooli-data-eng] componentize dbt
