+++ hooli-data-eng/hooli_data_eng/components/__init__.py
@@ -1,3 +1,4 @@
 from hooli_data_eng.components.scheduled_job_component import (
     ScheduledJobComponent as ScheduledJobComponent,
+    ScheduledPartitionedJobComponent as ScheduledPartitionedJobComponent,
 )
